Refactor texture programs
